[REF] move anf concept into separate header
