+++ ADOL-C/include/adolc/drivers/CMakeLists.txt
@@ -1,5 +1,6 @@
 install(FILES
         absnormalform.h
+        absnormalformconcept.h
         drivers.h
         odedrivers.h
         psdrivers.h
